🚚 Move config file in src
