+++ packages/server/src/db/db.ts
@@ -1,5 +1,5 @@
 import * as Knex from "knex";
-import { config } from "../../config";
+import { config } from "../config";
 
 export const knex = Knex({
   client: config.database_client,
