+++ packages/server/src/index.ts
@@ -2,7 +2,7 @@ import * as Koa from "koa";
 import * as cors from "@koa/cors";
 import { router } from "./router";
 import * as bodyParser from "koa-bodyparser";
-import { config } from "../config";
+import { config } from "./config";
 
 const app = new Koa();
